Added account app and migrated profile models
